fix sscanf types

+++ ext/native/ext/vjson/main.cpp
@@ -25,7 +25,7 @@ void populate_sources(const char *filter, std::vector<std::vector<char> > &sourc
 		}
 	}
 
-	printf("Loaded %d json files\n", sources.size());
+	printf("Loaded %d json files\n", (int)sources.size());
 }
 
 #define IDENT(n) for (int i = 0; i < n; ++i) printf("    ")
@@ -91,13 +91,13 @@ int main(int argc, char **argv)
 	int passed = 0;
 	for (size_t i = 0; i < sources.size(); ++i)
 	{
-		printf("Parsing %d\n", i + 1);
+		printf("Parsing %d\n", (int)i + 1);
 		if (parse(&sources[i][0]))
 		{
 			++passed;
 		}
 	}
-	printf("Passed %d from %d tests\n", passed, sources.size());
+	printf("Passed %d from %d tests\n", passed, (int)sources.size());
 
 	// Pass
 	sources.clear();
@@ -106,13 +106,13 @@ int main(int argc, char **argv)
 	passed = 0;
 	for (size_t i = 0; i < sources.size(); ++i)
 	{
-		printf("Parsing %d\n", i + 1);
+		printf("Parsing %d\n", (int)i + 1);
 		if (parse(&sources[i][0]))
 		{
 			++passed;
 		}
 	}
-	printf("Passed %d from %d tests\n", passed, sources.size());
+	printf("Passed %d from %d tests\n", passed, (int)sources.size());
 
 	return 0;
 }
